Partial vim testcase handler
The goal of this is to have a set of "vim unit tests" which non-vim text
editors that're adopting vim keybindings can adopt easily. The regular
vim unit tests aren't suitable for this, because (a) they're heavy on
complex vimscript, which a non-vim editor probably won't have, and (b)
they're too hard. So we have an easy-to-parse test case format, and a
vimscript program that'll execute them against real-vim as a reference
solution.

This commit starts the repository, adds half of the reference tester,
and provides a single example test case to demonstrate the syntax.