Better test runner, fixed level 3 tests
The previous version of the vimscript testrunner would parse files, and
try to run tests as it got to them. This didn't work very well, because
the best way to simulate keyboard input - feedkeys() - doesn't simulate
the keys immediately, but rather puts them in a queue, and simulates
them after returning all the way out. The old testrunner used :normal
instead of feedkeys, which sort-of worked but messed up the undo
history, and then the workaround for the undo history caused problems
with marks.  This commit mostly rewrites the testrunner to use feedkeys;
it first parses all the files, then each test calls feedkeys for the
keystrokes required to run it, followed by \<esc>:RunNextTest()\n.